requirements.txt

Django==5.0.6
djangorestframework==3.16.1
django-bootstrap==25.3
factory_boy==3.3.3
Faker==39.0.0
gunicorn


